refactor: :art: modify routers structure for testing
